move logic to scenario/conftest.py
